moved probability gate into a new gate impl folder
